+++ src/Interfaces/nodes.ts
@@ -1,3 +1,6 @@
+import { FontSizeName } from '@src/Interfaces/fonts'
+import { Elements } from '@src/Interfaces/elements'
+
 export type iconTypes =
   | '3d_rotation'
   | 'ac_unit'
@@ -942,3 +945,191 @@ export interface GridProperty {
   value: number
   unit: Units
 }
+
+export enum NodeTypes {
+  Root = 'Root',
+  Element = 'Element',
+  Component = 'Component',
+
+  Box = 'Box',
+  Text = 'Text',
+  Input = 'Input',
+  Image = 'Image',
+  Button = 'Button',
+  Link = 'Link',
+  Icon = 'Icon',
+}
+
+export enum Overflow {
+  visible = 'visible',
+  hidden = 'hidden',
+  scroll = 'scroll',
+}
+
+export enum Alignment {
+  stretch = 'stretch',
+  start = 'start',
+  center = 'center',
+  end = 'end',
+}
+
+export enum ObjectFit {
+  fill = 'fill',
+  contain = 'contain',
+  cover = 'cover',
+}
+
+export interface Padding {
+  top?: string
+  bottom?: string
+  left?: string
+  right?: string
+}
+
+interface SharedNodeProps {
+  id: string
+  alignment: {
+    horizontal: Alignment
+    vertical: Alignment
+  }
+  canBeDeleted?: boolean
+  canBeMoved?: boolean
+}
+
+type BorderPlaceholder = string
+type BoxShadowPlaceholder = string
+
+export interface RootNode extends SharedNodeProps {
+  type: NodeTypes.Root
+  nodeType: NodeTypes
+  children: Nodes[]
+  overflow: Overflow
+  boxShadow?: BoxShadowPlaceholder
+  padding: Padding
+  columns: GridProperty[]
+  rows: GridProperty[]
+  background: {
+    colorId: string
+  }
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+  border: BorderPlaceholder
+  focus: Partial<BoxNode>
+  hover: Partial<BoxNode>
+}
+
+export interface ElementNode extends SharedNodeProps {
+  type: NodeTypes.Element
+  elementType: keyof Elements
+  elementId: string
+  overrides: any
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+}
+
+export interface ComponentNode extends SharedNodeProps {
+  type: NodeTypes.Component
+  overrides: any
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+}
+
+export interface BoxNode extends SharedNodeProps {
+  type: NodeTypes.Box
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+  border: BorderPlaceholder
+  boxShadow?: BoxShadowPlaceholder
+  background: {
+    colorId: string
+  }
+  focus: Partial<BoxNode>
+  hover: Partial<BoxNode>
+}
+
+export interface TextNode extends SharedNodeProps {
+  type: NodeTypes.Text
+
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+  text: string
+  fontSize: FontSizeName
+  fontColorId?: string
+
+  focus: Partial<TextNode>
+  hover: Partial<TextNode>
+}
+
+export interface InputNode extends SharedNodeProps {
+  type: NodeTypes.Input
+
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+  background: {
+    colorId: string
+  }
+  border: BorderPlaceholder
+  boxShadow: BoxShadowPlaceholder
+  focus: Partial<InputNode>
+  hover: Partial<InputNode>
+}
+
+export interface ImageNode extends SharedNodeProps {
+  type: NodeTypes.Image
+  imageUrl: string
+  objectFit: ObjectFit
+
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+  //border: BorderPlaceholder
+  //boxShadow: BoxShadowPlaceholder
+  focus: Partial<ImageNode>
+  hover: Partial<ImageNode>
+}
+
+export interface IconNode extends SharedNodeProps {
+  type: NodeTypes.Icon
+  iconType: iconTypes
+
+  fontColorId?: string
+  position: {
+    columnStart: number
+    columnEnd: number
+    rowStart: number
+    rowEnd: number
+  }
+  //border: BorderPlaceholder
+  //boxShadow: BoxShadowPlaceholder
+  focus: Partial<ImageNode>
+  hover: Partial<ImageNode>
+}
+
+export type Nodes = RootNode | ElementNode | ComponentNode | TextNode | BoxNode | InputNode | ImageNode | IconNode
