+++ src/Interfaces/ui.ts
@@ -1,3 +1,7 @@
+import { Router } from '@src/Interfaces/router'
+import { Nodes, NodeTypes } from '@src/Interfaces/nodes'
+import { ComponentView } from '@src/Interfaces/components'
+
 export enum DragDirection {
   N = 'N',
   NE = 'NE',
@@ -12,4 +16,42 @@ export enum DragDirection {
 export enum ComponentStateMenu {
   hover = 'hover',
   focus = 'focus',
+}
+
+export interface AddingAtom {
+  type: NodeTypes
+  position: {
+    x: number
+    y: number
+  }
+  imageUrl?: string
+}
+
+export interface HoveredCell {
+  component: Nodes
+  rowIndex: number
+  colIndex: number
+}
+
+export interface UI {
+  router: Router
+  componentView: ComponentView
+  editingColorId: string
+  editingTextNode: Nodes
+  editingBoxNode: Nodes
+  addingComponent: boolean
+  showAddComponentMenu: boolean
+  selectedNode: Nodes
+  expandingNode: {
+    node: Nodes
+    parent: Nodes
+    direction: DragDirection
+  }
+  draggingNodePosition: {
+    x: number
+    y: number
+  }
+  addingAtom: AddingAtom
+  hoveredCell: HoveredCell
+  stateManager?: ComponentStateMenu
 }
