+++ src/editor/Components/Preview/AddComponentMenu/AddComponentMenu.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react'
 import styled from 'styled-components'
 import state from '@state'
-import { Alignment, NodeTypes, ObjectFit, Overflow } from '@src/interfaces'
+import { Alignment, NodeTypes, ObjectFit, Overflow, Units } from '@src/Interfaces/nodes'
 import { uuid } from '@src/utils'
 import { FontSizeName } from '@src/Interfaces/fonts'
-import { Units } from '@src/Interfaces/node'
 
 const Menu = styled.div`
   background: rgb(244, 244, 244);
