+++ src/editor/Components/Preview/AddingAtom.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import state from '@state'
 import { connect } from 'lape'
-import { NodeTypes } from '@src/interfaces'
+import { NodeTypes } from '@src/Interfaces/nodes'
 
 const AddingAtom = () => {
   if (state.ui.addingAtom.type === NodeTypes.Text) {
