+++ src/editor/Components/Preview/Preview.tsx
@@ -4,7 +4,7 @@ import state from '@state'
 import AddComponentMenu from './AddComponentMenu/AddComponentMenu'
 import Component from '@src/editor/Nodes/_Component'
 import { connect } from 'lape'
-import { ComponentView } from '@src/interfaces'
+import { ComponentView } from '@src/Interfaces/components'
 import AddingAtom from '@src/editor/Components/Preview/AddingAtom'
 import { getCurrentComponent } from '@src/selectors'
 import AddComponent from '@src/editor/TopBar/AddComponent'
