+++ src/editor/Nodes/_Component.tsx
@@ -6,12 +6,12 @@ import ImageAtom from './Image'
 import BoxAtom from './Box'
 import InputAtom from './Input'
 import TextAtom from './Text'
-import { Node, NodeTypes } from '@src/interfaces'
+import { Nodes, NodeTypes } from '@src/Interfaces/nodes'
 import IconAtom from '@src/editor/Nodes/Icon'
 
 interface Props {
-  component: Node
-  parent: Node | null
+  component: Nodes
+  parent: Nodes | null
 }
 function Component({ component, parent }: Props) {
   if (component.type === NodeTypes.Root) {
