+++ src/editor/Settings/colorList.ts
@@ -1,5 +1,5 @@
 import chroma from 'chroma-js'
-import { Color } from '@src/interfaces'
+import { Color } from '@src/Interfaces/state'
 
 type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
 type StaticColors = Omit<Color, 'id'>
